refactor: Move PartialSearchCube type to domain
